pyproject.toml
[tool.poetry]
name = "query_ai"
version = "1.0.0"
description = "A query application with AI"
authors = ["Ron Webb <ron@ronella.xyz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
transformers = "^4.49.0"
torch = "^2.6.0"
pgvector = "^0.3.6"
psycopg2 = "^2.9.10"
nltk = "^3.9.1"
flask = "^3.1.0"
python-dotenv = "^1.0.1"
waitress = "^3.0.2"


[tool.poetry.group.dev.dependencies]
cython = "^3.0.11"
pytest = "^8.3.4"
pylint = "^3.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
lint = "query_ai.util.lint_runner:main"
app = "query_ai.application:main"